move uncertain prediction to predict
